% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday}
\usage{
birthday(x)
}
\arguments{
\item{x}{amount of people}
}
\value{
The probability that at least two people share a birthday in a group of x people
}
\description{
Birthday
}
\examples{
birthday(20:25)
}
