% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set.}
\format{
A data frame with 144 rows and 6 variables:
\describe{
  \item{RIVER}{name of river}
  \item{MILE}{mile marker along the river}
  \item{SPECIES}{type of fish}
  \item{LENGTH}{length in cm}
  \item{WEIGHT}{weight of the fish, in gms}
  \item{DDT}{value of DDT in ppm}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
ddt
}
\description{
A data set collected by the US forces from the Tennessee river and its tributaries
}
\keyword{datasets}
