% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIRE data set.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance from fire station to fire, in miles}
  \item{DAMAGE}{Thousands of dollars damage}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set from MS that shows data from fire damage incidents.
}
\keyword{datasets}
