% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Sampling Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Sampling size}

\item{p}{Probability of success}
}
\value{
A Binomial bar plot showing the binomial probability for the number of successes.
}
\description{
Sampling Function
}
\examples{
mybin(iter=10000, n=10, p=0.7)
}
