% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{CLT Binomial}
\usage{
mycltb(n, iter, p = 0.5)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{p}{probability}
}
\value{
A CLT Binomial distribution
}
\description{
CLT Binomial
}
\examples{
\dontrun{mycltb(n=10, iter=10000, p=0.3)}
}
