% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{A numeric vector representing the sample data.}

\item{mu}{A numeric vector of mean values to evaluate.}

\item{sig}{A numeric vector of standard deviation values to evaluate.}

\item{...}{Additional arguments passed to the contour function.}
}
\value{
A list with components 'x', 'coord', and 'maxl', and a contour plot of the likelihood function using a normal distribution.
}
\description{
mymlnorm
}
\examples{
\dontrun{mymlnorm(x=c(10,12,13,15,12,11,10),mu=seq(10,15,length=1000),
sig=seq(0.1,4,length=1000),lwd=2,labcex=1)}
}
