% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of seats available on the flight}

\item{gamma}{The probability that overbooking will occur}

\item{p}{The likelihood that a passenger will show up for the flight}
}
\value{
Returns the plots for both discrete and continuous methods, and a list of values for nd, nc, N, p, and gamma
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N = 400, gamma = 0.02, p = 0.95)}
}
