% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce.R
\docType{data}
\name{spruce}
\alias{spruce}
\title{SPRUCE data set.}
\format{
A data frame with 36 rows and 2 variables:
\describe{
  \item{BHDiameter}{breast height diameter in cm}
  \item{Height}{height in m}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
spruce
}
\description{
A data set from MS that measures different components of spruce trees.
}
\keyword{datasets}
