% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{value to compute P(X <= a)}
}
\value{
A named list with components 'mu', 'sigma', 'a', and 'area' (the cumulative probability P(X <= a)).
}
\description{
myncurve
}
\examples{
\dontrun{myncurve(mu = 10, sigma = 5, a = 6)}
}
